Added Logic to Place Order

+++ SpringEcommerce/src/main/java/org/example/springecommerce/model/dto/OrderItemResponse.java
@@ -1,10 +1,10 @@
 package org.example.springecommerce.model.dto;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public record OrderItemResponse(
         String name,
         int quantity,
-        BigInteger totalPrice
+        BigDecimal totalPrice
 ) {
 }
